requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-environ==0.12.0
psycopg2-binary==2.9.10
drf-yasg==1.21.10
django-cors-headers==4.7.0
django-filter==25.1
Pillow==10.4.0 